Splitted the codes
